+++ src/hdmf/common/table.py
@@ -292,9 +292,15 @@ class DynamicTable(Container):
             {'name': 'colnames', 'type': 'array_data',
              'doc': 'the ordered names of the columns in this table. columns must also be provided.',
              'default': None},
+            {'name': 'target_tables',
+             'doc': ('dict mapping DynamicTableRegion column name to the table that the DTR points to. The column is '
+                     'added to the table if it is not already present (i.e., when it is optional).'),
+             'type': dict,
+             'default': None},
             allow_positional=AllowPositional.WARNING)
     def __init__(self, **kwargs):  # noqa: C901
         id, columns, desc, colnames = popargs('id', 'columns', 'description', 'colnames', kwargs)
+        target_tables = popargs('target_tables', kwargs)
         super().__init__(**kwargs)
         self.description = desc
 
@@ -468,6 +474,10 @@ class DynamicTable(Container):
         self.__colids = {name: i + 1 for i, name in enumerate(self.colnames)}
         self._init_class_columns()
 
+        if target_tables:
+            self._set_dtr_targets(target_tables)
+
+
     def __set_table_attr(self, col):
         if hasattr(self, col.name) and col.name not in self.__uninit_cols:
             msg = ("An attribute '%s' already exists on %s '%s' so this column cannot be accessed as an attribute, "
@@ -515,6 +525,40 @@ class DynamicTable(Container):
                     if col.get('enum', False):
                         self.__uninit_cols[col['name'] + '_elements'] = col
                         setattr(self, col['name'] + '_elements', None)
+
+    def _set_dtr_targets(self, target_tables: dict):
+        """Set the target tables for DynamicTableRegion columns.
+
+        If a column is not yet initialized, it is initialized with the target table.
+        """
+        for colname, table in target_tables.items():
+            if colname not in self:  # column has not yet been added (it is optional)
+                column_conf = None
+                for conf in self.__columns__:
+                    if conf['name'] == colname:
+                        column_conf = conf
+                        break
+                if column_conf is None:
+                    raise ValueError("'%s' is not the name of a predefined column of table %s."
+                                        % (colname, self))
+                if not column_conf.get('table', False):
+                    raise ValueError("Column '%s' must be a DynamicTableRegion to have a target table."
+                                        % colname)
+                self.add_column(name=column_conf['name'],
+                                description=column_conf['description'],
+                                index=column_conf.get('index', False),
+                                table=True)
+            if isinstance(self[colname], VectorIndex):
+                col = self[colname].target
+            else:
+                col = self[colname]
+            if not isinstance(col, DynamicTableRegion):
+                raise ValueError("Column '%s' must be a DynamicTableRegion to have a target table." % colname)
+            # if columns are passed in, then the "table" attribute may have already been set
+            if col.table is not None and col.table is not table:
+                raise ValueError("Column '%s' already has a target table that is not the passed table." % colname)
+            if col.table is None:
+                col.table = table
 
     @staticmethod
     def __build_columns(columns, df=None):